tests/wpt/tests/html/canvas/element/reset/2d.reset.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.reset.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.reset.basic</h1>
<p class="desc">reset clears to transparent black</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("reset clears to transparent black");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, ctx.canvas.width, ctx.canvas.height);
  ctx.reset();
  _assertPixel(canvas, 0,0, 0,0,0,0);
  _assertPixel(canvas, 50,25, 0,0,0,0);
  _assertPixel(canvas, 25,50, 0,0,0,0);
  _assertPixel(canvas, 100,50, 0,0,0,0);
  _assertPixel(canvas, 0,50, 0,0,0,0);
  _assertPixel(canvas, 100,0, 0,0,0,0);
  t.done();

});
</script>
